+++ packages/riverpod/test/framework/async_value_test.dart
@@ -31,7 +31,7 @@ void main() {
       const AsyncValue.data(42).when(
         data: (value) => [value],
         error: (a, b) => throw Error(),
-        loading: () => throw Error(),
+        loading: (_) => throw Error(),
       ),
       [42],
     );
@@ -42,18 +42,20 @@ void main() {
       AsyncValue<int>.error(42, stack).when(
         data: (value) => throw Error(),
         error: (a, b) => [a, b],
-        loading: () => throw Error(),
+        loading: (_) => throw Error(),
       ),
       [42, stack],
     );
 
     expect(
-      const AsyncValue<int>.loading().when(
+      const AsyncValue<int>.loading(
+        previous: AsyncData(42),
+      ).when(
         data: (value) => throw Error(),
         error: (a, b) => throw Error(),
-        loading: () => 'loading',
+        loading: (previous) => 'loading ${previous?.value}',
       ),
-      'loading',
+      'loading 42',
     );
   });
 
@@ -78,11 +80,13 @@ void main() {
       );
 
       expect(
-        const AsyncValue<int>.loading().maybeWhen(
-          loading: () => 'loading',
+        const AsyncValue<int>.loading(
+          previous: AsyncData(42),
+        ).maybeWhen(
+          loading: (previous) => 'loading ${previous?.value}',
           orElse: () => throw Error(),
         ),
-        'loading',
+        'loading 42',
       );
     });
 
@@ -90,7 +94,7 @@ void main() {
       expect(
         const AsyncValue.data(42).maybeWhen(
           error: (a, b) => throw Error(),
-          loading: () => throw Error(),
+          loading: (_) => throw Error(),
           orElse: () => 'orElse',
         ),
         'orElse',
@@ -101,7 +105,7 @@ void main() {
       expect(
         AsyncValue<int>.error(42, stack).maybeWhen(
           data: (value) => throw Error(),
-          loading: () => throw Error(),
+          loading: (_) => throw Error(),
           orElse: () => 'orElse',
         ),
         'orElse',
@@ -137,10 +141,12 @@ void main() {
       );
 
       expect(
-        const AsyncValue<int>.loading().whenOrNull(
-          loading: () => 'loading',
+        const AsyncValue<int>.loading(
+          previous: AsyncData(42),
+        ).whenOrNull(
+          loading: (previous) => 'loading ${previous?.value}',
         ),
-        'loading',
+        'loading 42',
       );
     });
 
@@ -148,7 +154,7 @@ void main() {
       expect(
         const AsyncValue.data(42).whenOrNull(
           error: (a, b) => throw Error(),
-          loading: () => throw Error(),
+          loading: (_) => throw Error(),
         ),
         null,
       );
@@ -158,7 +164,7 @@ void main() {
       expect(
         AsyncValue<int>.error(42, stack).whenOrNull(
           data: (value) => throw Error(),
-          loading: () => throw Error(),
+          loading: (_) => throw Error(),
         ),
         null,
       );
@@ -252,30 +258,30 @@ void main() {
     );
 
     expect(
-      const AsyncValue<int>.loading().toString(),
-      'AsyncLoading<int>()',
+      const AsyncValue<int>.loading(previous: AsyncData(42)).toString(),
+      'AsyncLoading<int>(previous: AsyncData<int>(value: 42))',
     );
   });
 
   test('AsyncValue.whenData', () {
     expect(
       const AsyncValue.data(42).whenData((value) => '$value'),
-      isA<AsyncData<String>>().having((s) => s.value, 'value', '42'),
+      const AsyncData<String>('42'),
     );
     expect(
       const AsyncValue<int>.loading().whenData((value) => '$value'),
-      isA<AsyncLoading<String>>(),
+      const AsyncLoading<String>(),
     );
     expect(
       const AsyncValue<int>.error(21).whenData((value) => '$value'),
-      isA<AsyncError<String>>().having((s) => s.error, 'error', 21),
+      const AsyncError<String>(21),
     );
   });
 
   test('AsyncValue.asData', () {
     expect(
       const AsyncValue.data(42).asData,
-      isA<AsyncData<int>>().having((s) => s.value, 'value', 42),
+      const AsyncData<int>(42),
     );
     expect(const AsyncValue<void>.loading().asData, isNull);
     expect(AsyncValue<void>.error(Error()).asData, isNull);
@@ -285,7 +291,7 @@ void main() {
     expect(const AsyncValue.data(42).value, 42);
     expect(
       () => const AsyncValue<void>.loading().value,
-      throwsA(isA<AsyncValueLoadingException>()),
+      throwsA(isA<AsyncValueLoadingError>()),
     );
 
     final error = Error();
@@ -298,7 +304,7 @@ void main() {
   test('AsyncValue.data handles null', () {
     expect(
       const AsyncValue<int?>.data(null).data,
-      isA<AsyncData<int?>>().having((s) => s.value, 'value', null),
+      const AsyncData<int?>(null),
     );
   });